+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - feat: add default batching to `predict` (#229)
+- BREAKING CHANGE: drop Python 3.7 support (#250)
 
 ### Changed