+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- feat: `_local_predict` has been re added (#248)
 - feat: add default batching to `predict` (#229)
 - BREAKING CHANGE: drop Python 3.7 support (#250)